Implement stop functionality
This keeps track of which tasks have been stopped in a global hash and
uses a monitor interval to exit when they're all stopped. This doesn't
work well in the case where the task has ended but a large amount of
time remains in the interval before the task runs again. In order to
provide more immediate stopping, the concurrent_count will need to
move into the global hash so it can be checked by the monitor.