Create task state database
This allows both the concurrent count and the task mode (running,
paused or stopping) to be accessible so that it can be periodically
read from a control file or altered via other means, such as on
reception of a signal.

It also moves a bunch of logic out of the futures callback functions,
making those easier to understand.