Bugfix for starting on the hour or minute with additional delay
Previously, if the requested start hour had additional minutes and
that minute had not been reached in the current hour, the start time
would be in the next hour, not the current one. The upshot was that
the first interval was skipped.

The same bug applied to start minutes with additional seconds.

Now, periodic only skips to the next interval if the requested
additional delay (minute or second) has already occurred in the
current interval.